+++ rpg unity/Assets/character/skill/Magician/magic floor/MagicFloor.cs
@@ -16,9 +16,9 @@ public class MagicFloor : MonoBehaviourPunCallbacks
     private float dealIncreasePerSkillLevel = 1;
     private float dealIncreasePerPower = 1;
 
-    private int flatHeal = 1;
-    private int healIncreasePerSkillLevel = 1;
-    private int healIncreasePerPower = 1;
+    private float flatHeal = 1;
+    private float healIncreasePerSkillLevel = 1;
+    private float healIncreasePerPower = 1;
 
     private float duration = 3f;
